+++ android/src/main/java/com/lodev09/truesheet/TrueSheetView.kt
@@ -11,11 +11,11 @@ import com.facebook.react.uimanager.ThemedReactContext
 import com.facebook.react.uimanager.UIManagerHelper
 import com.facebook.react.uimanager.events.EventDispatcher
 import com.google.android.material.bottomsheet.BottomSheetBehavior
-import com.lodev09.truesheet.core.DismissEvent
-import com.lodev09.truesheet.core.MountEvent
-import com.lodev09.truesheet.core.PresentEvent
 import com.lodev09.truesheet.core.RootSheetView
-import com.lodev09.truesheet.core.SizeChangeEvent
+import com.lodev09.truesheet.events.DismissEvent
+import com.lodev09.truesheet.events.MountEvent
+import com.lodev09.truesheet.events.PresentEvent
+import com.lodev09.truesheet.events.SizeChangeEvent
 
 class TrueSheetView(context: Context) :
   ViewGroup(context),
